requirements.txt
asgiref==3.2.10
beautifulsoup4==4.9.3
certifi==2020.6.20
confusable-homoglyphs==3.2.0
Django==3.1.3
django-bootstrap4==2.3.1
django-crispy-forms==1.9.2
django-tinymce==3.1.0
Pillow==8.0.1
psycopg2-binary==2.8.6
pytz==2020.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.4.1
urllib3==1.25.11
